Extract ParameterPositionTemplate
Extract `ParameterPositionTemplate` into separate compilation unit to allow granular use in annotation feature template.
